Rails 3.2.6 & Redmine 2.0 compatibility fixes
The plugin seems to work without routes, so get rid of them because the
old one are not compatible with Rails 3.

The link in the menu triggers a crash and I couldn't fix it. Because
it's useless to us, remove it completely.

Finally, we add a symlink to http_auth_helper.rb in lib/ because Ruby
can't find it in app/helpers.

NOTE: I know nothing about Ruby & Rails.